move config methods to another file
