+++ config.py
@@ -1,0 +1,1 @@
+multi_column_editor/config.py
